feat: 媒体经纬度

+++ database/migrations/2023_02_27_134449_create_short_videos_table.php
@@ -20,8 +20,8 @@ class CreateShortVideosTable extends Migration
             $table->string('title')->comment('视频标题');
             $table->string('cover')->comment('封面');
             $table->string('video_url')->comment('视频地址');
-            $table->float('longitude')->default(0)->comment('经度');
-            $table->float('latitude')->default(0)->comment('纬度');
+            $table->decimal('longitude', 9, 6)->default(0)->comment('经度');
+            $table->decimal('latitude', 8, 6)->default(0)->comment('纬度');
             $table->string('address')->default('')->comment('具体地址');
             $table->integer('is_private')->default(0)->comment('是否为私密视频：0-否，1-是');
             $table->integer('like_number')->default(0)->comment('点赞数');
